requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.5
